refactor: use `typeof` in `HFTEST_ASSERT_OP`
The `__typeof` keyword can be used instead of a union to declare
variables with unknown type. This simplifies the implementation of
`HFTEST_ASSERT`, and ensures that enums are not silently casted to ints.

The only drawback compared to the old method is that `typeof` cannot be
used on bitfields. But bitfields are rare, so we can fix all the uses by
loading the bitfield into a local variable first, then comparing it.

+++ test/vmapi/ffa_secure_partition_only/secure_partition.c
@@ -174,9 +174,10 @@ TEST(ffa_features, succeeds_ffa_call_ids)
 	ret = ffa_features(FFA_RXTX_MAP_64);
 	EXPECT_EQ(ret.func, FFA_SUCCESS_32);
 	rxtx_map_params = ffa_features_rxtx_map_params(ret);
-	EXPECT_EQ(rxtx_map_params.min_buf_size, FFA_RXTX_MAP_MIN_BUF_4K);
-	EXPECT_EQ(rxtx_map_params.mbz, 0);
-	EXPECT_EQ(rxtx_map_params.max_buf_size,
+	EXPECT_EQ((uint8_t)rxtx_map_params.min_buf_size,
+		  FFA_RXTX_MAP_MIN_BUF_4K);
+	EXPECT_EQ((uint16_t)rxtx_map_params.mbz, 0);
+	EXPECT_EQ((uint16_t)rxtx_map_params.max_buf_size,
 		  FFA_RXTX_MAP_MAX_BUF_PAGE_COUNT);
 #endif
 }
